move StatusModal to components/base
